feat: type별 컴포넌트 분기
Resolve: #249
